% Generated by roxygen2 (4.0.2): do not edit by hand
\name{normal}
\alias{normal}
\title{A Normalization tool of the association data}
\usage{
normal(df)
}
\arguments{
\item{df}{}
}
\value{
return the normalized value of the imagery data
}
\description{
This function normalizes the data first
}
\author{
Abhishek Kapoor
}

