% Generated by roxygen2 (4.0.2): do not edit by hand
\name{SolarMaps}
\alias{SolarMaps}
\title{A graphical tool with better visualization of association data}
\usage{
SolarMaps(data)
}
\arguments{
\item{df}{data frame for brand association}
}
\value{
data frame with classification in each column about Positive and Negative
differentation
}
\description{
This function normalizes the data produced by the
}
\details{
This function return the positive and negative differentated imagery attributes
}
\author{
Abhishek Kapoor
}
\seealso{
\code{normal}
}

